+++ packages/seda-ui/src/index.ts
@@ -10,3 +10,4 @@ export * from "./components/Radio";
 export * from "./components/Toggle";
 export * from "./components/Badge";
 export * from "./components/Tag";
+export * from "./components/Alert";